refactor: Update button styles in DonationSection and PostSection components
Update the button styles in the DonationSection and PostSection components to improve the visual appearance and user experience.

In DonationSection.style.css, change the background color of the .donate-button:hover class to #b42c7c.

In PostSection.css, add a font-weight of 500 to the .view-all-button class.

In PostSection.jsx, update the text of the view all button to 'הצגת הכל'